added settings screen + implemented singleton on npcspawnvariables

+++ Assets/Scripts/BaseClassCharacter.cs
@@ -11,7 +11,7 @@ public class BaseClassCharacter : MonoBehaviour
     public bool Attack { get; set; }
     public bool Hostile { get; set; }
 
-    int health = 300;
+    float health = 300;
 
     public BaseClassCharacter()
     {
@@ -19,13 +19,13 @@ public class BaseClassCharacter : MonoBehaviour
 
     }
 
-    public void getPunched(int enemy_damage)
+    public void getPunched(float enemy_damage)
     {
         health -= enemy_damage;
 
     }
 
-    public int getHealth()
+    public float getHealth()
     {
         return health;
